added vscode settings, fixed vars.json, moved files around
