+++ .gitignore
@@ -8,3 +8,4 @@
 **/.DS_Store
 
 ./src/commands/bootcamp
+/config/dev/vars.json
