+++ .vscode/launch.json
@@ -1,0 +1,21 @@
+{
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "type": "node",
+      "request": "launch",
+      "name": "Run Bot Locally",
+      "skipFiles": [
+        "<node_internals>/**"
+      ],
+      "preLaunchTask": "buildBot",
+      "runtimeExecutable": "npm",
+      "runtimeArgs": [
+          "run-script", "local:run"
+      ],
+      "outFiles": [
+        "${workspaceFolder}/dist/**/*.js"
+      ]
+    }
+  ]
+}
