+++ .vscode/tasks.json
@@ -1,0 +1,13 @@
+{
+  "version": "2.0.0",
+  "tasks": [
+    {
+      "type": "npm",
+      "script": "ts:build",
+      "group": "build",
+      "problemMatcher": [],
+      "label": "buildBot",
+      "detail": "tsc"
+    }
+  ]
+}
