+++ src/commandDetails/admin/setTimer.ts
@@ -1,11 +1,16 @@
-import { CategoryChannel, GuildMember, Permissions, VoiceChannel } from "discord.js";
+import {
+  CategoryChannel,
+  GuildMember,
+  Permissions,
+  VoiceChannel,
+} from "discord.js";
 import {
   CodeyCommandDetails,
   CodeyCommandOptionType,
   SapphireMessageExecuteType,
   SapphireMessageResponse,
 } from "../../codeyCommand";
-import { lowestInt } from "../../commands/bootcamp/utils";
+import { lowestInt } from "../../utils/bootcamp";
 import { BootcampSettings } from "../../components/bootcamp";
 
 const setTimerExecuteCommand: SapphireMessageExecuteType = async (
